Float the quota

+++ lib/private/legacy/OC_Helper.php
@@ -577,7 +577,7 @@ class OC_Helper {
 	/**
 	 * Get storage info including all mount points and quota
 	 */
-	private static function getGlobalStorageInfo(int $quota, IUser $user, IMountPoint $mount): array {
+	private static function getGlobalStorageInfo(float $quota, IUser $user, IMountPoint $mount): array {
 		$rootInfo = \OC\Files\Filesystem::getFileInfo('', 'ext');
 		$used = $rootInfo['size'];
 		if ($used < 0) {
